Module Protocol_version_header

This library offers a lightweight way for applications protocols to version themselves. The more protocols that add themselves to Known_protocol, the nicer error messages we will get when connecting to a service while using the wrong protocol.

type t

include Core_kernel.Bin_prot.Binable.S with type t := t

include Core_kernel.Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create_exn : protocol:Protocol_version_header__Known_protocol.t -> supported_versions:int list -> t

create_exn ~protocol ~supported_version raises if List.length supported_versions >= 100

val negotiate : allow_legacy_peer:bool -> us:t -> peer:t -> int Core_kernel.Or_error.t

negotiate ~allow_legacy_peer ~us ~peer inspects the magic numbers of us and peer. If the magic numbers match, the highest shared version number is returned.

If allow_legacy_peer then the magic number of peer is assumed to be us if no magic number exists.

val contains_magic_prefix : protocol:Protocol_version_header__Known_protocol.t -> bool Core_kernel.Bin_prot.Type_class.reader

contains_magic_prefix reads a bin_protted value of type t and returns a boolean saying whether this magic number was observed.

val any_magic_prefix : Protocol_version_header__Known_protocol.t option Core_kernel.Bin_prot.Type_class.reader

any_magic_prefix reads the magic number for one of the known protocols.

module For_test : sig ... end